Estamina fase 2
La estamina ya baja y sube bien, para correr y no correr, e incluso recibir daño (con la tecla -)

+++ Game/FinalProject/Assets/Scripts/StaminaBar.cs
@@ -7,12 +7,12 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxStamina(int stamina){
+    public void SetMaxStamina(float stamina){
         slider.maxValue = stamina;
         slider.value = stamina;
     }
 
-    public void SetStamina(int stamina){
+    public void SetStamina(float stamina){
         slider.value = stamina;
     }
 }
